pyproject.toml
[tool.poetry]
name = "miio2prometheus"
version = "0.1.0"
description = ""
authors = ["Maxim Andreev <andreevmaxim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-miio = "^0.5.4"
prometheus-client = "^0.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
